Fix time out setting and testing
Two problems were noticed. The Java methods seem to have been renamed in OmniDriver. Add functions
using the new names with "USB" and changed the old functions to use setTimeout if available but fall back on setUSBTimeout if setTimeout is missing. A private function was implemented for the test of methods' availability.